more util restructure
